pyproject.toml
[project]
name = "cordis-pic-ror"
version = "0.1.0"
description = "Resolve CORDIS participant PIC identifiers to ROR identifiers from CORDIS, OpenAIRE, ROR, and Zenodo community sources."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.0",
]

[dependency-groups]
dev = []
